fastapi/requirements.txt
absl-py==2.3.1
accelerate==1.1.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
bitsandbytes==0.48.2
certifi==2026.2.25
cfgv==3.4.0
charset-normalizer==3.4.5
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
datasets==4.5.0
dill==0.4.0
distlib==0.4.0
efficientnet-pytorch==0.7.1
einops==0.8.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
facexlib==0.3.0
fastapi==0.128.8
filelock==3.19.1
filterpy==1.4.5
flatbuffers==25.12.19
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
hdbscan==0.8.41
hf-xet==1.4.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
icecream==2.1.10
identify==2.6.15
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.10.0
kiwisolver==1.4.7
lazy-loader==0.5
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.7.5
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
namex==0.1.0
networkx==3.2.1
nodeenv==1.10.0
numba==0.60.0
numpy==2.0.2
openai-clip==1.0.1
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiqa==0.1.14.1
pyparsing==3.3.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-discovery==1.1.3
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
ruff==0.15.5
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
starlette==0.49.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.25
tokenizers==0.15.2
tomli==2.4.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
transformers==4.37.2
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
virtualenv==21.2.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.6
wrapt==2.1.2
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
